test_vertex_ai_streaming_response_id
