+++ tests/local_testing/test_amazing_vertex_completion.py
@@ -3839,3 +3839,103 @@ def test_vertex_schema_test():
     print(response)
 
 
+
+
+def test_vertex_ai_response_id():
+    """Test that litellm preserves the response ID from Vertex AI's API for non-streaming responses"""
+    from litellm.llms.custom_httpx.http_handler import HTTPHandler
+    load_vertex_ai_credentials()
+
+    client = HTTPHandler()
+
+    mock_response = MagicMock()
+    mock_response.status_code = 200
+    mock_response.headers = {"Content-Type": "application/json"}
+    mock_response.json.return_value = {
+        "responseId": "vertex_ai_response_123",
+        "candidates": [
+            {
+                "content": {
+                    "role": "model",
+                    "parts": [{"text": "Hello! How can I help you today?"}],
+                },
+                "finishReason": "STOP",
+                "safetyRatings": [
+                    {
+                        "category": "HARM_CATEGORY_HATE_SPEECH",
+                        "probability": "NEGLIGIBLE",
+                    }
+                ],
+            }
+        ],
+        "usageMetadata": {
+            "promptTokenCount": 10,
+            "candidatesTokenCount": 8,
+            "totalTokenCount": 18,
+        },
+    }
+
+    with patch.object(client, "post", return_value=mock_response) as mock_post:
+        response = completion(
+            model="vertex_ai/gemini-1.5-pro",
+            messages=[{"role": "user", "content": "Hi!"}],
+            client=client,
+        )
+
+        # Verify the response ID is preserved
+        assert response.id == "vertex_ai_response_123"
+        assert response.choices[0].message.content == "Hello! How can I help you today?"
+
+
+def test_vertex_ai_streaming_response_id():
+    """Test that litellm preserves the response ID from Vertex AI's API for streaming responses"""
+    from litellm.llms.custom_httpx.http_handler import HTTPHandler
+    from litellm.llms.vertex_ai.gemini.vertex_and_google_ai_studio_gemini import (
+        make_sync_call,
+    )
+    load_vertex_ai_credentials()
+
+    client = HTTPHandler()
+
+    def mock_post(url, **kwargs):
+        def stream_response():
+            chunk = {
+                "responseId": "vertex_ai_response_stream_123",
+                "candidates": [
+                    {
+                        "content": {
+                            "role": "model",
+                            "parts": [{"text": "Hello streaming!"}],
+                        },
+                        "finishReason": "STOP",
+                    }
+                ],
+                "usageMetadata": {
+                    "promptTokenCount": 10,
+                    "candidatesTokenCount": 8,
+                    "totalTokenCount": 18,
+                },
+            }
+            yield json.dumps(chunk)
+
+        mock_response = MagicMock()
+        mock_response.status_code = 200
+        mock_response.iter_lines = MagicMock(return_value=stream_response())
+        return mock_response
+
+    logging_obj = MagicMock()
+
+    with patch.object(client, "post", side_effect=mock_post):
+        iterator = make_sync_call(
+            client=client,
+            gemini_client=None,
+            api_base="https://mock-vertex-ai-api.com",
+            headers={},
+            data="{}",
+            model="gemini-pro",
+            messages=[],
+            logging_obj=logging_obj,
+        )
+        iterator = iter(iterator)
+        first_chunk = next(iterator)
+        assert first_chunk.id == "vertex_ai_response_stream_123"
